ui/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pydantic>=2.0.0
pyyaml>=6.0.1
plotly>=5.17.0
pandas>=2.1.0
streamlit-chat>=0.1.1
streamlit-ace>=0.1.1 